Add OWL RL reasoner stage and fix ontology consistency bugs
- Add scripts/reasoner.py: runs an OWL 2 RL reasoner (owlrl), validates
  consistency, and materializes DL inferences (union, intersection, and
  transitive class memberships) into rdf/elden_ring_reasoned.ttl
- Fix disjointness violations at the source in converter.py: boss drop
  lists contained location headers (trailing ':') and 'Other Drops'
  separators that were emitted as drops, forcing Locations to be typed
  as Items
- Disambiguate entities that collide across disjoint families, such as
  the Fire Knight Queelign NPC versus the Spirit Ash of the same name
- Change attack and weight ranges from xsd:float to xsd:decimal so exact
  game stats pass strict OWL RL datatype validation
- Regenerate all rdf/*.ttl outputs; ignore the krr2 virtualenv
- Document the new reasoner stage in the README

+++ scripts/schema.py
@@ -151,7 +151,9 @@ def build_comprehensive_ontology():
     # Weight is unique per item.
     g.add((ER.weight, RDF.type, OWL.DatatypeProperty))
     g.add((ER.weight, RDF.type, OWL.FunctionalProperty))
-    g.add((ER.weight, RDFS.range, XSD.float))
+    # xsd:decimal, not xsd:float: game stats are exact values, and OWL RL's strict
+    # datatype checker rejects lexical forms like "0.0" as xsd:float.
+    g.add((ER.weight, RDFS.range, XSD.decimal))
 
     # Inverse Functional Property (Requirement: "At least one... inverse functional")
     # The Game ID uniquely identifies the item (if we had it), or we can treat the Name as a unique key.
@@ -175,7 +177,7 @@ def build_comprehensive_ontology():
     for dt in dmg_types:
         uri = ER[f"attack{dt}"]
         g.add((uri, RDF.type, OWL.DatatypeProperty))
-        g.add((uri, RDFS.range, XSD.float))
+        g.add((uri, RDFS.range, XSD.decimal))
 
     for attr in ["Strength", "Dexterity", "Intelligence", "Faith", "Arcane"]:
         g.add((ER[f"requires{attr}"], RDF.type, OWL.DatatypeProperty))
